fixed routes and wrote readme

+++ Develop/routes/api/product-routes.js
@@ -1,44 +1,46 @@
-const router = require("express").Router();
-const { Product, Category, Tag, ProductTag } = require("../../models");
+const router = require('express').Router();
+const { Product, Category, Tag, ProductTag } = require('../../models');
 
 // The `/api/products` endpoint
 
 // get all products
-router.get("/", async (req, res) => {
+router.get('/', async (req, res) => {
   // find all products
   // be sure to include its associated Category and Tag data
   try {
-    const productData = await Product.findAll({
-      include: [{ model: Tag, through: ProductTag }, { model: Category }],
+    const allProductsData = await Product.findAll({
+      include: [
+        { model: Category }, 
+        { model: Tag, through: ProductTag },
+      ],
     });
-    res.status(200).json(productData);
+    res.status(200).json(allProductsData);
+
   } catch (err) {
-    res.status(400).json(err);
+    res.status(500).json(err);
   }
 });
 
 // get one product
-router.get("/:id", async (req, res) => {
+router.get('/:id', async (req, res) => {
   // find a single product by its `id`
   // be sure to include its associated Category and Tag data
   try {
-    const productData = await Product.findByPk(req.params.id, {
-      include: [{ model: Tag, through: ProductTag }, { model: Category }],
+    const oneProductData = await Product.findByPk(req.params.id, {
+      include: [
+        { model: Category }, 
+        { model: Tag, through: ProductTag },
+      ],
     });
+    res.status(200).json(oneProductData);
 
-    if (!productData) {
-      res.status(404).json({ message: "No product found with this id!" });
-      return;
-    }
-
-    res.status(200).json(productData);
   } catch (err) {
     res.status(500).json(err);
   }
 });
 
 // create new product
-router.post("/", (req, res) => {
+router.post('/', (req, res) => {
   /* req.body should look like this...
     {
       product_name: "Basketball",
@@ -70,7 +72,7 @@ router.post("/", (req, res) => {
 });
 
 // update product
-router.put("/:id", (req, res) => {
+router.put('/:id', (req, res) => {
   // update product data
   Product.update(req.body, {
     where: {
@@ -78,57 +80,65 @@ router.put("/:id", (req, res) => {
     },
   })
     .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+      if (req.body.tagIds && req.body.tagIds.length) {
+        
+        ProductTag.findAll({
+          where: { product_id: req.params.id }
+        }).then((productTags) => {
+          // create filtered list of new tag_ids
+          const productTagIds = productTags.map(({ tag_id }) => tag_id);
+          const newProductTags = req.body.tagIds
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              product_id: req.params.id,
+              tag_id,
+            };
+          });
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+            // figure out which ones to remove
+          const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
+                  // run both actions
+          return Promise.all([
+            ProductTag.destroy({ where: { id: productTagsToRemove } }),
+            ProductTag.bulkCreate(newProductTags),
+          ]);
+        });
+      }
+
+      return res.json(product);
     })
-    .then((updatedProductTags) => res.json(updatedProductTags))
     .catch((err) => {
       // console.log(err);
       res.status(400).json(err);
     });
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete('/:id', async (req, res) => {
   // delete one product by its `id` value
   try {
-    const productData = await Product.destroy({
-      where: {
-        id: req.params.id,
-      },
-    });
+    const deleteProductData = await Product.findByPk(req.params.id);
 
-    if (!productData) {
-      res.status(404).json({ message: "No product found with that id!" });
+    if (!deleteProductData) {
+      res.status(404).json({ error: 'Unable to find product' });
       return;
     }
 
-    res.status(200).json(productData);
+    await Product.destroy({
+      where: {
+        id: req.params.id
+      }
+    });
+
+    res.status(200).json(deleteProductData);
   } catch (err) {
+    
+    console.error(err);
     res.status(500).json(err);
   }
 });
+
 
 module.exports = router;
